Update ProgressDialog Components and create ProgressDialog Composable

+++ app/src/main/java/domilopment/apkextractor/data/ProgressDialogUiState.kt
@@ -3,7 +3,7 @@ package domilopment.apkextractor.data
 data class ProgressDialogUiState(
     val title: String? = null,
     val process: String? = null,
-    val progress: Int = 0,
+    val progress: Float = 0f,
     val tasks: Int = 0,
     val shouldBeShown: Boolean = false
 )
